Move modalcontent,trigger and modal to slatevideo
