+++ src/components/SlateEditor/plugins/video/SlateVideo.tsx
@@ -8,17 +8,22 @@
 
 import { useCallback, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Editor, Element, Transforms } from "slate";
+import { Editor, Element, Path, Transforms } from "slate";
 import { ReactEditor, RenderElementProps, useSelected } from "slate-react";
 import styled from "@emotion/styled";
+import { IconButtonV2 } from "@ndla/button";
 import { colors } from "@ndla/core";
+import { Pencil } from "@ndla/icons/action";
 import { DeleteForever, Link } from "@ndla/icons/editor";
+import { Modal, ModalContent, ModalTrigger } from "@ndla/modal";
 import { SafeLinkIconButton } from "@ndla/safelink";
 import { BrightcoveMetaData } from "@ndla/types-embed";
 import { BrightcoveEmbed } from "@ndla/ui";
-import EditVideo from "./EditVideo";
+import EditVideo, { FormValues } from "./EditVideo";
 import { BrightcoveEmbedElement, TYPE_EMBED_BRIGHTCOVE } from "./types";
 import { useBrightcoveMeta } from "../../../../modules/embed/queries";
+import { inlineContentToHTML } from "../../../../util/articleContentConverter";
+import { addBrightCoveTimeStampVideoid } from "../../../../util/videoUtil";
 import Spinner from "../../../Spinner";
 import { StyledDeleteEmbedButton, StyledFigureButtons } from "../embed/FigureButtons";
 
@@ -45,6 +50,7 @@ interface Props extends RenderElementProps {
 
 const SlateVideo = ({ attributes, element, editor, children }: Props) => {
   const [hasError, setHasError] = useState(false);
+  const [open, setOpen] = useState(false);
 
   const { t, i18n } = useTranslation();
 
@@ -72,35 +78,82 @@ const SlateVideo = ({ attributes, element, editor, children }: Props) => {
     [brightcoveQuery.data, brightcoveQuery.error, element.data],
   );
 
+  const onClose = () => {
+    setOpen(false);
+    ReactEditor.focus(editor);
+    const path = ReactEditor.findPath(editor, element);
+    if (Editor.hasPath(editor, Path.next(path))) {
+      setTimeout(() => {
+        Transforms.select(editor, Path.next(path));
+      }, 0);
+    }
+  };
+
+  const onSave = (values: FormValues) => {
+    Transforms.setNodes(
+      editor,
+      {
+        data: {
+          ...embed?.embedData,
+          alt: values.alttext,
+          caption: inlineContentToHTML(values.caption),
+          videoid: addBrightCoveTimeStampVideoid(values.videoid, values.startTime),
+        },
+      },
+      {
+        match: (node) => Element.isElement(node) && node.type === TYPE_EMBED_BRIGHTCOVE,
+        at: ReactEditor.findPath(editor, element),
+      },
+    );
+    onClose();
+  };
+
   return (
-    <VideoWrapper {...attributes} data-selected={isSelected} data-error={hasError} contentEditable={false}>
-      <StyledFigureButtons>
-        {embed && (
-          <>
-            <EditVideo embed={embed} editor={editor} element={element} setHasError={setHasError} />
-            <SafeLinkIconButton
-              title={t("form.video.brightcove")}
-              aria-label={t("form.video.brightcove")}
-              colorTheme="light"
-              to={`https://studio.brightcove.com/products/videocloud/media/videos/${embed.embedData.videoid}`}
-            >
-              <Link />
-            </SafeLinkIconButton>
-          </>
-        )}
-        <StyledDeleteEmbedButton
-          aria-label={t("form.video.remove")}
-          title={t("form.video.remove")}
-          colorTheme="danger"
-          onClick={removeVideo}
-          data-testid="remove-video-element"
-        >
-          <DeleteForever />
-        </StyledDeleteEmbedButton>
-      </StyledFigureButtons>
-      {!embed || brightcoveQuery.isLoading ? <Spinner /> : <BrightcoveEmbed embed={embed} />}
-      {children}
-    </VideoWrapper>
+    <>
+      <Modal open={open} onOpenChange={setOpen}>
+        <VideoWrapper {...attributes} data-selected={isSelected} data-error={hasError} contentEditable={false}>
+          {!embed ? (
+            <Spinner />
+          ) : (
+            <StyledFigureButtons>
+              <ModalTrigger>
+                <IconButtonV2
+                  aria-label={t("form.video.editVideo")}
+                  title={t("form.video.editVideo")}
+                  colorTheme="light"
+                >
+                  <Pencil />
+                </IconButtonV2>
+              </ModalTrigger>
+              <SafeLinkIconButton
+                title={t("form.video.brightcove")}
+                aria-label={t("form.video.brightcove")}
+                colorTheme="light"
+                to={`https://studio.brightcove.com/products/videocloud/media/videos/${embed.embedData.videoid}`}
+              >
+                <Link />
+              </SafeLinkIconButton>
+              <StyledDeleteEmbedButton
+                aria-label={t("form.video.remove")}
+                title={t("form.video.remove")}
+                colorTheme="danger"
+                onClick={removeVideo}
+                data-testid="remove-video-element"
+              >
+                <DeleteForever />
+              </StyledDeleteEmbedButton>
+            </StyledFigureButtons>
+          )}
+          <ModalContent>
+            {element.data && (
+              <EditVideo onClose={onClose} onSave={onSave} embed={element.data} setHasError={setHasError} />
+            )}
+          </ModalContent>
+          {!embed || brightcoveQuery.isLoading ? <Spinner /> : <BrightcoveEmbed embed={embed} />}
+          {children}
+        </VideoWrapper>
+      </Modal>
+    </>
   );
 };
